+++ src/main/java/org/software/testing/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.software.testing;
 
 public class Main {
     public static void main(String[] args) {
