+++ src/test/java/org/software/testing/MathUtilsTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.software.testing;
 
 import org.junit.jupiter.api.Test;
